Rename FileUploadGraphQLView to GraphQLView
