+++ saleor/graphql/core/mutations.py
@@ -11,9 +11,8 @@ from graphql_jwt.exceptions import JSONWebTokenError, PermissionDenied
 
 from ...account import models
 from ..account.types import User
-from ..file_upload.types import Upload
 from ..utils import get_nodes
-from .types.common import Error
+from .types import Error, Upload
 from .utils import snake_to_camel_case
 
 registry = get_global_registry()
